requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.12
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
exceptiongroup==1.0.4
idna==3.4
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
pytest-asyncio==0.20.2
pytz==2022.6
requests==2.28.1
shortuuid==1.0.11
simplejson==3.18.0
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.12
